requirements.txt
pdf2image==1.17.0
pytesseract==0.3.10
opencv-python==4.9.0.80
numpy>=1.24.0,<2.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytest==7.4.3
Pillow>=9.0.0
python-dotenv==1.0.0